requirements.txt
agency-swarm==0.4.4
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.6.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
datamodel-code-generator==0.26.1
deepdiff==6.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
genson==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.8.2
jsonref==1.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
openai==1.65.4
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.7.1
sniffio==1.3.1
termcolor==2.4.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
mem0ai==0.1.0
